Move cached to common
