+++ controllers/apply/form-router-next/index.js
@@ -21,7 +21,7 @@ const {
 const commonLogger = require('../../../common/logger');
 const appData = require('../../../common/appData');
 const { localify } = require('../../../common/urls');
-const { noStore } = require('../../../middleware/cached');
+const { noStore } = require('../../../common/cached');
 const { requireActiveUserWithCallback } = require('../../../middleware/authed');
 const { injectCopy } = require('../../../middleware/inject-content');
