+++ controllers/user/index.js
@@ -6,7 +6,7 @@ const moment = require('moment');
 
 const { Users } = require('../../db/models');
 const { localify, redirectForLocale } = require('../../common/urls');
-const { noStore } = require('../../middleware/cached');
+const { noStore } = require('../../common/cached');
 const { requireNotStaffAuth } = require('../../middleware/authed');
 const { injectCopy } = require('../../middleware/inject-content');
