+++ controllers/user/update-email.js
@@ -5,7 +5,7 @@ const express = require('express');
 
 const { Users } = require('../../db/models');
 const { redirectForLocale } = require('../../common/urls');
-const { csrfProtection } = require('../../middleware/cached');
+const { csrfProtection } = require('../../common/cached');
 const { requireUserAuth } = require('../../middleware/authed');
 const { injectCopy } = require('../../middleware/inject-content');
 
@@ -26,11 +26,7 @@ function render(req, res, data = null, errors = []) {
 }
 
 async function handleSubmission(req, res, next) {
-
-    const validationResult = validateSchema(
-        emailOnly(req.i18n),
-        req.body
-    );
+    const validationResult = validateSchema(emailOnly(req.i18n), req.body);
 
     if (validationResult.isValid) {
         try {
